+++ src/components/icons/LeafIcon.tsx
@@ -1,0 +1,23 @@
+interface IProps {
+  width: number
+  height: number
+  className?: string
+  onClick?: () => void
+}
+
+export default function LeafIcon(props: IProps) {
+  const {width, height, className, onClick} = props
+
+  return (
+    <div className={className} onClick={onClick}>
+      <svg xmlns="http://www.w3.org/2000/svg" className="icon icon-tabler icon-tabler-leaf" width={width} height={height}
+           viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" fill="none" strokeLinecap="round"
+           strokeLinejoin="round">
+        <path stroke="none" d="M0 0h24v24H0z" fill="none"/>
+        <path d="M5 21c.5 -4.5 2.5 -8 7 -10"/>
+        <path d="M9 18c6.218 0 10.5 -3.288 11 -12v-2h-4.014c-9 0 -11.986 4 -12 9c0 1 0 3 2 5h3z"/>
+      </svg>
+    </div>
+  )
+}
+